+++ src/main/java/rsystems/commands/Say.java
@@ -22,7 +22,7 @@ public class Say extends ListenerAdapter{
 
         String[] args = event.getMessage().getContentRaw().split("\\s+");
         //Sponge command
-        /*
+
         if (args[0].equalsIgnoreCase(HiveBot.prefix + HiveBot.commands.get(28).getCommand())) {
             if(RoleCheck.getRank(event,event.getMember().getId()) >= HiveBot.commands.get(28).getRank()) {
                 LOGGER.info(HiveBot.commands.get(28).getCommand() + " called by " + event.getAuthor().getAsTag());
@@ -71,7 +71,7 @@ public class Say extends ListenerAdapter{
                             .queue();
                 }
             }
-        }*/
+        }
     }
 
 }