+++ changelog.md
@@ -1,7 +1,11 @@
 # Change Log
+
 ### 0.17.3
 - Changed referenceTriggers to send direct message unless -show used
 - Updated who command with more information
+- Added class for SQLite DB Handling
+  - Added method to add privileged users to db automatically.
+- Added AutoRemove class to remove privileged roles from user for setting inactive status.
 
 ### 0.17.2
 - Changed info and reference triggers to allow to be called in private messages.